コメント削除修正

+++ app/controllers/comments_controller.rb
@@ -10,9 +10,9 @@ class CommentsController < ApplicationController
   def destroy
     comment = Comment.find(params[:id])
     @practice = Practice.find(params[:practice_id])
-    if user_signed_in? && @practice.comment.user_id == current_user.id
+    if user_signed_in? && comment.practice_id == @practice.id && comment.user_id == current_user.id
       comment.destroy
-      redirect_to practice_path(@practice.id)
+      redirect_to practice_path(comment.practice.id)
     end
   end
